+++ sysarmy-sueldos-simulacion/src/Form/Form.css
@@ -1,0 +1,4 @@
+.form-main-container {
+  margin-top: 70px;
+  padding-top: 10px;
+}
