+++ sysarmy-sueldos-simulacion/src/Navbar/Navbar.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+import './Navbar.css'
+
+class Navbar extends Component {
+  render() {
+    return (
+      <div className="navbar">
+        
+      </div>
+    )
+  }
+}
+
+export default Navbar
